+++ cdk/lib/network-stack.mjs
@@ -1,7 +1,4 @@
 import { Stack, Duration } from 'aws-cdk-lib';
-import iam from 'aws-cdk-lib/aws-iam';
-import cdk from 'aws-cdk-lib';
-import rds from 'aws-cdk-lib/aws-rds';
 import ec2 from 'aws-cdk-lib/aws-ec2';
 
 export class NetworkStack extends Stack {
@@ -17,6 +14,21 @@ export class NetworkStack extends Stack {
         this.vpc = new ec2.Vpc(this, props.formatName('vpc'), {
             maxAzs: 2,
             ipAddresses: ec2.IpAddresses.cidr('10.0.0.0/16'),
+            enableDnsHostnames: true,
+            enableDnsSupport: true,
+            natGateways: 0,
+            subnetConfiguration: [
+                {
+                    cidrMask: 22,
+                    name: props.formatName('public'),
+                    subnetType: ec2.SubnetType.PUBLIC, // for web access 
+                },
+                {
+                    cidrMask: 22,
+                    name: props.formatName('private'),
+                    subnetType: ec2.SubnetType.PRIVATE_WITH_EGRESS, // for database access
+                },
+            ],
         });
 
     }