+++ ving/docs/change-log.md
@@ -5,6 +5,10 @@ outline: deep
 
 ## August 2024
 
+### 2024-08-19
+* Started working on the web stack for CDK.
+* Updated the database stack to use the secrets manager for the database credentials.
+
 ### 2024-08-09
 * Started replacing Pulumi with CDK.
 * Added a CDK stack for uploads.